refactored audio visualizer into own component
