+++ src/components/SpliceVideo/VideoControls.tsx
@@ -1,54 +1,13 @@
 import { FC, useEffect, useState, useMemo } from 'react'
 import useSpliceVideo from '@/hooks/useSpliceVideo'
-import { Box, IconButton, Slider, Stack, Tooltip, Typography } from '@mui/material'
+import { Box, IconButton, Stack, Tooltip } from '@mui/material'
 import { FirstPage, LastPage, PlayArrow, SkipNext, SkipPrevious, Pause, Replay, Replay5, Forward5, Replay10, Forward10 } from '@mui/icons-material'
 import fs from 'fs'
-import WavesurferPlayer from '@wavesurfer/react'
-import SpectrogramPlugin from "wavesurfer.js/dist/plugins/spectrogram.js"
-import TimelinePlugin from 'wavesurfer.js/dist/plugins/timeline.js'
-import colormap from 'colormap'
-import { ipcRenderer } from 'electron'
-
-const colors = colormap({
-  colormap: 'hot',
-  nshades: 256,
-  format: 'float',
-  alpha: 1,
-})
 
 const VideoControls: FC = () => {
-  const { selectedVideo, videoFramerate, updateVideoRef } = useSpliceVideo()
-
-  const video = document.getElementById('splice-video') as HTMLVideoElement
+  const { selectedVideo, videoFramerate, updateVideoRef, videoRef } = useSpliceVideo()
 
   const [videoState, setVideoState] = useState<'playing' | 'paused'>('paused')
-
-  const [zoom, setZoom] = useState<number>(1)
-  const [frequencyMax, setFrequencyMax] = useState<number>(22_050)
-  const [audioSampleRate, setAudioSampleRate] = useState<number>(44_100)
-
-  useEffect(() => {
-    if (!selectedVideo) {
-      return
-    }
-
-    // restore values to default
-    setZoom(1)
-    setFrequencyMax(22_050)
-    setAudioSampleRate(44_100)
-
-    ipcRenderer.send('get-audio-sample-rate', {
-      filePath: selectedVideo,
-    })
-
-    ipcRenderer.once('got-audio-sample-rate', (_, sampleRate) => {
-      setAudioSampleRate(sampleRate)
-    })
-  }, [selectedVideo])
-
-  useEffect(() => {
-    setFrequencyMax(audioSampleRate / 2)
-  }, [audioSampleRate])
 
   const videoUrl = useMemo(() => {
     if (!selectedVideo) {
@@ -60,7 +19,7 @@ const VideoControls: FC = () => {
   }, [selectedVideo])
 
   useEffect(() => {
-    if (!video) {
+    if (!videoRef) {
       return
     }
 
@@ -72,28 +31,28 @@ const VideoControls: FC = () => {
       setVideoState('paused')
     }
 
-    video.addEventListener('play', handlePlay)
-    video.addEventListener('pause', handlePause)
+    videoRef.addEventListener('play', handlePlay)
+    videoRef.addEventListener('pause', handlePause)
 
     return () => {
-      video.removeEventListener('play', handlePlay)
-      video.removeEventListener('pause', handlePause)
+      videoRef.removeEventListener('play', handlePlay)
+      videoRef.removeEventListener('pause', handlePause)
     }
-  }, [video])
+  }, [videoRef])
 
   const handleGoToStart = () => {
-    video.currentTime = 0
+    videoRef!.currentTime = 0
   }
 
   const handleGoToEnd = () => {
-    video.currentTime = video.duration
+    videoRef!.currentTime = videoRef!.duration
   }
 
   const handlePlayPause = () => {
-    if (video.paused) {
-      video.play()
+    if (videoRef!.paused) {
+      videoRef!.play()
     } else {
-      video.pause()
+      videoRef!.pause()
     }
   }
 
@@ -102,7 +61,7 @@ const VideoControls: FC = () => {
       return
     }
 
-    video.currentTime -= 1 / videoFramerate
+    videoRef!.currentTime -= 1 / videoFramerate
   }
 
   const handleGoForwardOneFrame = () => {
@@ -110,11 +69,11 @@ const VideoControls: FC = () => {
       return
     }
 
-    video.currentTime += 1 / videoFramerate
+    videoRef!.currentTime += 1 / videoFramerate
   }
 
   const handleSecondsOffset = (seconds: number) => {
-    video.currentTime += seconds
+    videoRef!.currentTime += seconds
   }
 
   return (
@@ -211,67 +170,6 @@ const VideoControls: FC = () => {
                 </Tooltip>
               </Box>
             </Stack>
-          </Box>
-          <Box>
-            <WavesurferPlayer
-              height={100}
-              media={video}
-              progressColor='#1976d2'
-              minPxPerSec={zoom}
-              dragToSeek
-              normalize
-              // @ts-ignore
-              splitChannels
-              sampleRate={audioSampleRate}
-              frequencyMax={frequencyMax}
-              onRedraw={(wavesurfer) => {
-                const activePlugins = wavesurfer.getActivePlugins()
-
-                // detect if timeline plugin is active
-                if (!activePlugins.some((plugin) => (plugin as any).timelineWrapper)) {
-                  wavesurfer.registerPlugin(TimelinePlugin.create())
-                }
-
-                // detect if spectrogram plugin is active
-                if (!activePlugins.some((plugin) => (plugin as any).colorMap)) {
-                  wavesurfer.registerPlugin(SpectrogramPlugin.create({
-                    labels: true,
-                    frequencyMax,
-                    labelsBackground: '#00000066',
-                    colorMap: colors,
-                    splitChannels: false,
-                  }))
-                }
-              }}
-            />
-            <Box maxWidth='calc(100% - 32px)'>
-              <Stack direction='row' alignItems='center' gap={2}>
-                <Typography variant='body2'>
-                  Zoom
-                </Typography>
-                <Slider
-                  value={zoom}
-                  onChange={(_, value) => setZoom(value as number)}
-                  min={1}
-                  max={5000}
-                />
-              </Stack>
-              <Stack direction='row' alignItems='center' gap={2}>
-                <Typography variant='body2' noWrap>
-                  Max Frequency (Hz)
-                </Typography>
-                <Slider
-                  value={frequencyMax}
-                  onChange={(_, value) => setFrequencyMax(value as number)}
-                  min={0}
-                  max={audioSampleRate / 2}
-                  valueLabelDisplay='auto'
-                  sx={{
-                    flex: 1,
-                  }}
-                />
-              </Stack>
-            </Box>
           </Box>
         </>
       )}
