+++ src/pages/SpliceVideo.tsx
@@ -1,5 +1,5 @@
 import useSpliceVideo from '@/hooks/useSpliceVideo'
-import { Box, Button, Typography, Stack, Grid, IconButton } from '@mui/material'
+import { Box, Button, Typography, Stack, Grid } from '@mui/material'
 import { FC, useEffect } from 'react'
 import { ipcRenderer } from 'electron'
 import Progress from '@/components/SpliceVideo/Progress'
@@ -8,6 +8,7 @@ import VideoControls from '@/components/SpliceVideo/VideoControls'
 import AutoSplice from '@/components/SpliceVideo/AutoSplice'
 import SaveProject from '@/components/SpliceVideo/SaveProject'
 import LoadProject from '@/components/SpliceVideo/LoadProject'
+import AudioVisualizers from '@/components/SpliceVideo/AudioVisualizers'
 
 const SpliceVideo: FC = () => {
   const { updateSelectedVideo, selectedVideo } = useSpliceVideo()
@@ -59,6 +60,7 @@ const SpliceVideo: FC = () => {
               selectedVideo &&
               <Stack spacing={2}>
                 <VideoControls />
+                <AudioVisualizers />
                 <AutoSplice />
               </Stack>
             }
